move models into app directory
